Fix RPC endpoint calls (#226)
RPC handlers were accidentally dropped during the session refactoring. Fixes #223
