+++ electrumx_rpc
@@ -34,7 +34,6 @@ simple_commands = {
 
 session_commands = {
     "disconnect": "Disconnect sessions",
-    "log": "Control logging of sessions",
 }
 
 other_commands = {
@@ -132,6 +131,8 @@ def main():
     if port is None:
         port = int(environ.get("RPC_PORT", 8000))
     method = args.pop("command")
+    if not method:
+        raise Exception("Must specify a method")
     timeout = args.pop("timeout")
 
     # aiorpcX makes this so easy...
